# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Work Agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/settings/agent": {
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Update Agent",
        "operationId": "update_agent_api_v1_settings_agent_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/openai": {
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Openai Settings",
        "operationId": "openai_settings_api_v1_settings_openai_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeySettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/microsoft": {
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Microsoft Settings",
        "operationId": "microsoft_settings_api_v1_settings_microsoft_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicrosoftSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/microsoft/auth": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Microsoft Auth",
        "operationId": "microsoft_auth_api_v1_microsoft_auth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUrl"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/microsoft/discover/{kind}": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Discover",
        "operationId": "discover_api_v1_microsoft_discover__kind__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent"
            }
          },
          {
            "name": "site_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryEntry"
                  },
                  "title": "Response Discover Api V1 Microsoft Discover  Kind  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Sources",
        "operationId": "sources_api_v1_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceView"
                  },
                  "type": "array",
                  "title": "Response Sources Api V1 Sources Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Create Source",
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}": {
      "patch": {
        "tags": [
          "connections"
        ],
        "summary": "Update Source",
        "operationId": "update_source_api_v1_sources__source_id__patch",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/sync": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Sync",
        "operationId": "sync_api_v1_sources__source_id__sync_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/github/connections": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "summary": "Connections",
        "operationId": "connections_api_v1_github_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionView"
                  },
                  "type": "array",
                  "title": "Response Connections Api V1 Github Connections Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Create Connection",
        "operationId": "create_connection_api_v1_github_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/connections/{identity}": {
      "patch": {
        "tags": [
          "GitHub"
        ],
        "summary": "Update Connection",
        "operationId": "update_connection_api_v1_github_connections__identity__patch",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentSettings": {
        "properties": {
          "model": {
            "type": "string",
            "maxLength": 200,
            "title": "Model",
            "default": ""
          },
          "embedding_model": {
            "type": "string",
            "maxLength": 200,
            "title": "Embedding Model",
            "default": "text-embedding-3-small"
          },
          "instructions": {
            "type": "string",
            "maxLength": 15000,
            "title": "Instructions"
          },
          "style": {
            "type": "string",
            "maxLength": 2000,
            "title": "Style"
          },
          "daily_limit": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Daily Limit"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "instructions",
          "style",
          "daily_limit",
          "paused",
          "allowed_actions"
        ],
        "title": "AgentSettings"
      },
...
      "AuthStart": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "write": {
            "type": "boolean",
            "title": "Write",
            "default": false
          }
        },
        "type": "object",
        "title": "AuthStart"
      },
...
      "AuthUrl": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AuthUrl"
      },
...
      "ConnectionInput": {
        "properties": {
          "owner": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9-]{0,38}$",
            "title": "Owner"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "token": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "owner",
          "name",
          "token"
        ],
        "title": "ConnectionInput"
      },
...
      "ConnectionUpdate": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ConnectionUpdate"
      },
      "ConnectionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "last_discovery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Discovery"
          },
          "retry_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner",
          "name",
          "enabled",
          "demo",
          "status",
          "error",
          "last_discovery",
          "retry_at"
        ],
        "title": "ConnectionView"
      },
...
      "DiscoveryEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DiscoveryEntry"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "run_after": {
            "type": "string",
            "format": "date-time",
            "title": "Run After"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "payload",
          "status",
          "attempts",
          "run_after",
          "created_at"
        ],
        "title": "JobView"
      },
      "KeySettings": {
        "properties": {
          "api_key": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "KeySettings"
      },
...
      "MicrosoftSettings": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "maxLength": 2000,
            "title": "Client Secret",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "client_id"
        ],
        "title": "MicrosoftSettings"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "Ok"
      },
...
      "SourceConfiguration": {
        "properties": {
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 90
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          },
          "list_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Id"
          },
          "drive_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Id"
          },
          "repository_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id"
          },
          "connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SourceConfiguration"
      },
      "SourceInput": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Name"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled",
            "default": false
          },
          "writable": {
            "type": "boolean",
            "title": "Writable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "title": "SourceInput"
      },
...
      "SourceUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "ai_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Enabled"
          },
          "writable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Writable"
          },
          "days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3650.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Days"
          }
        },
        "type": "object",
        "title": "SourceUpdate"
      },
      "SourceView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "config": {
            "$ref": "#/components/schemas/SourceConfiguration"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "writable": {
            "type": "boolean",
            "title": "Writable"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cursor": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cursor"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "next_sync": {
            "type": "string",
            "format": "date-time",
            "title": "Next Sync"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "name",
          "config",
          "enabled",
          "ai_enabled",
          "writable",
          "status",
          "cursor",
          "next_sync",
          "created_at"
        ],
        "title": "SourceView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}